e_reader/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
gcloud==0.17.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
iniconfig==1.1.1
jws==0.1.3
oauth2client==3.0.0
packaging==21.3
pluggy==1.0.0
protobuf==4.21.10
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==3.0.9
Pyrebase==3.0.27
pytest==7.2.0
python-jwt==2.0.1
pytz==2022.6
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
